+++ dart/attributes/lib/data.dart
@@ -20,4 +20,3 @@ export 'src/data.dart' hide DataObjectView, DataArrayView;
 
 // export also `Identifier` - essential for using data objects and arrays.
 export 'src/values.dart' show Identifier;
-